Removed Raw Storage functions
Neither one of these functions were needed since session now has store
attribute and an fs WriteStream is only created when the attribute is
set to true. This eliminates use for any FS storage if the raw path is
not provided and the resulting email will only be stored in the database
and never on FileSystem.