[fix] SpamcPassThrough Kept Connection alive
Spamc check was set to go off after Mailparser had fired an 'end' event
but the stream was not consumed until MailParser had ended. This caused
an issue with large emails as MailParser stream did not fire the 'end'
event. The alternative is now set to require a stored version of email
(raw) in order to be piped into spamc after database storage is done.
This equally reduces strain on SMTP connection but may result in slower
processing and mail status checks but fortunately the Model has a
placeholder for status.